requirements.txt
Django==1.5.4
django-bootstrap3==1.1.0
django-endless-pagination==2.0
django-markdown-deux==1.0.4
markdown2==2.1.0
wsgiref==0.1.2
